Restructured project
Now all asset groups and resource groups have their own module.
Jobs are defined in a separate file.
